Fix test flakiness by avoiding RF method recursion.
When a method under test takes as a parameter a reference to its own
class (or if there's a circular path of multiple such methods and
classes), it is possible for generated RamFuzz code to exceed the
recursion depth limit while processing a top-level method invoked from
main().  In that case, the method's side-effects will not take place,
and main() will detect it as a failure.  This happens rarely (requiring
`depthlimit` roulette spins with particular results) but it does happen;
it's better to eliminate it and have the tests always pass.