Ensure template names are always qualified.
Previously, a template name would not be explicitly qualified when processed by
TypePrinter (eg, as an argument to another template).  That screwed up the
declaration of std::string parameters.

Now rfstream prints the template name and makes sure it's always qualified.  One
consequence of this is that rfstream no longer prints the ElaboratedType
qualifier, to avoid printing it twice; instead, this is left to the desugared
type.

Signed-off-by: Dejan Mircevski <github@mircevski.com>