Replace RamFuzz::inheritance with a finish() arg.
This is in preparation for running both RamFuzz and InheritanceBuilder
in the same ClangTool run.  The Inheritance object will not be complete
until after the tool has run.

Remove findInheritance(ClangTool), as it's dead now.

Remove genTests(string), as nobody calls it, and it's incorrect
anyway (doesn't invoke RamFuzz::finish()).

Signed-off-by: Dejan Mircevski <github@mircevski.com>