Remove mutation functionality.
It is not safe.  When a part of the replay is skipped, some random
values are replaced by different ones.  This can result in different
control flow once the replay resumes, leading to misinterpretation of
the resumed input log.  For instance, an int in the log may be
interpreted as a float by the code, because an earlier mutated value has
a far-reaching effect that causes the int generation to be bypassed.

While regions theoretically may be able to prevent this, that's only
true if they are arbitrarily large.

Signed-off-by: Dejan Mircevski <github@mircevski.com>