Reduce type tag to only three values
We'll cast all integers to int64_t, all unsigned integers to uint64_t,
and all floats to double.  We'll generate random values in those three
types, send them back, and cast them to the original type in RamFuzz
runtime.  We already had to do this for types like char and unsigned
int, which zmqpp doesn't natively support.  It's simpler to do it for
nearly all types.

It might be possible to simplify further and only use one type, avoiding
type tags altogether.  But it's not obvious how to do that without
losing precision.

Signed-off-by: Dejan Mircevski <github@mircevski.com>