Adapt test/test.py for valgen
The script test.py runs end-to-end tests: it processes input code with
ramfuzz, compiles the output, and runs the resulting executable.  This
was fine when RamFuzz runtime was self-contained, but now it depends on
valgen running in the background.  And most tests really just want to
ensure the ramfuzz output compiles and links correctly; running the
executable is non-essential.  With this change, we turn off the running
of executables.  (In the future, we will add the ability to optionally
run the executable, when such tests become necessary.)

Because the RamFuzz runtime no longer replays logs, we delete all tests
for logging.

Signed-off-by: Dejan Mircevski <github@mircevski.com>